+++ frontend/pages/manufacturer-models/_manufacturerModelId/platforms.vue
@@ -633,6 +633,7 @@ export default class ManufacturerModelPlatformsSearchPage extends Vue {
       types: [],
       permissionGroups: [],
       onlyOwnPlatforms: false,
+      onlySelfCreatedPlatforms: false,
       includeArchivedPlatforms: false,
       manufacturerName: this.manufacturerModel!.manufacturerName,
       model: this.manufacturerModel!.model