+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Added:
 - [UFZ] Replaced Bitnami-ElasticSearch Images with standard upstream ES-Image (bumped Helm-Chart) ([Merge Request](https://codebase.helmholtz.cloud/hub-terra/sms/orchestration/-/merge_requests/687))
 - Deploy the FZJ SMS staging instance on the new VM ([Merge Request](https://codebase.helmholtz.cloud/hub-terra/sms/orchestration/-/merge_requests/684)) 
 
+Changed:
+- Extended search seperates between entries where the user was the creator and those where the user is a contact for ([Merge Request](https://codebase.helmholtz.cloud/hub-terra/orchestration/-/merge_requests/690))
+
 Fixed:
 - Resolved the MC make bucket issue for FZJ ([Merge Request](https://codebase.helmholtz.cloud/hub-terra/sms/orchestration/-/merge_requests/680))